+++ CHANGELOG.md
@@ -56,6 +56,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - 通貨転送をChevre口座取引で実装
 - メンバーシップオファー承認時のアプリケーションクライアントに対する排他処理調整
 - @cinerino/sdkで再構築
+- AccountTypeの使用を廃止してCurrencyTypeに変更
 
 ### Deprecated